Simpler implementation with Alure.
Still some stuff need to be done with VoiceSE. For now VOICE is entirely disabled.
There are some pops and clicks at the end of sounds, Alure loading bug?

Also, there is some stuff to do when "ch->entchannel == CHAN_STREAM" on StartStatic and StartDynamic at the end of the function. A dummy file path buffer is being used.